+++ packages/common/src/graphql/types.ts
@@ -76,4 +76,5 @@ export enum FieldScalar {
   String = 'String',
   Date = 'Date',
   Boolean = 'Boolean',
+  Bytes = 'Bytes',
 }