+++ packages/node/src/indexer/api.service.spec.ts
@@ -3,7 +3,7 @@
 
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { ApiPromise, WsProvider } from '@polkadot/api';
-import { ProjectNetwork } from '@subql/common/project/models';
+import { ProjectManifestVersioned, ProjectNetworkV0_0_1 } from '@subql/common';
 import { omit } from 'lodash';
 import { SubqueryProject } from '../configure/project.model';
 import { ApiService } from './api.service';
@@ -20,7 +20,7 @@ jest.mock('@polkadot/api', () => {
   return { ApiPromise, WsProvider: jest.fn() };
 });
 
-const testNetwork: ProjectNetwork = {
+const testNetwork: ProjectNetworkV0_0_1 = {
   endpoint: 'wss://kusama.api.onfinality.io/public-ws',
   types: {
     TestType: 'u32',
@@ -45,8 +45,14 @@ const testNetwork: ProjectNetwork = {
 };
 
 function testSubqueryProject(): SubqueryProject {
-  const project = new SubqueryProject();
-  project.network = testNetwork;
+  const project = new SubqueryProject(
+    new ProjectManifestVersioned({
+      specVersion: '0.0.1',
+      network: testNetwork,
+      dataSources: [],
+    } as any),
+    '',
+  );
   return project;
 }