+++ packages/node/src/indexer/api.service.test.ts
@@ -5,19 +5,29 @@ import { INestApplication } from '@nestjs/common';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { Test } from '@nestjs/testing';
 import { BlockHash } from '@polkadot/types/interfaces';
+import { ProjectManifestVersioned } from '@subql/common';
 import { take } from 'rxjs/operators';
 import { SubqueryProject } from '../configure/project.model';
 import { delay } from '../utils/promise';
 import { ApiService } from './api.service';
 
-function testSubqueryProject(): SubqueryProject {
-  const project = new SubqueryProject();
-  project.network = {
-    endpoint: 'wss://kusama.api.onfinality.io/public-ws',
-    types: {
-      TestType: 'u32',
-    },
-  };
+const WS_ENDPOINT = 'wss://kusama.api.onfinality.io/public-ws';
+const HTTP_ENDPOINT = 'https://kusama.api.onfinality.io/public';
+
+function testSubqueryProject(endpoint: string): SubqueryProject {
+  const project = new SubqueryProject(
+    new ProjectManifestVersioned({
+      specVersion: '0.0.1',
+      network: {
+        endpoint,
+        types: {
+          TestType: 'u32',
+        },
+      },
+      dataSources: [],
+    } as any),
+    '',
+  );
   return project;
 }
 
@@ -29,10 +39,15 @@ describe('ApiService', () => {
     return app?.close();
   });
 
-  const prepareApiService = async (): Promise<ApiService> => {
+  const prepareApiService = async (
+    endpoint: string = WS_ENDPOINT,
+  ): Promise<ApiService> => {
     const module = await Test.createTestingModule({
       providers: [
-        { provide: SubqueryProject, useFactory: testSubqueryProject },
+        {
+          provide: SubqueryProject,
+          useFactory: () => testSubqueryProject(endpoint),
+        },
         ApiService,
       ],
       imports: [EventEmitterModule.forRoot()],
@@ -359,4 +374,13 @@ describe('ApiService', () => {
     expect(expectedValidators1).toEqual(vs1);
     expect(expectedValidators2).toEqual(vs2);
   }, 300000);
+
+  it('support http provider', async () => {
+    const apiService = await prepareApiService(HTTP_ENDPOINT);
+    const api = apiService.getApi();
+    const patchedApi = await apiService.getPatchedApi();
+    const blockhash = await api.rpc.chain.getBlockHash(1);
+    await apiService.setBlockhash(blockhash, true);
+    await expect(patchedApi.query.system.events()).resolves.toHaveLength(2);
+  }, 30000);
 });