+++ packages/node/src/meta/health.service.ts
@@ -3,12 +3,15 @@
 
 import { Injectable } from '@nestjs/common';
 import { OnEvent } from '@nestjs/event-emitter';
+import { NodeConfig } from '../configure/NodeConfig';
 import {
   IndexerEvent,
   NetworkMetadataPayload,
   ProcessBlockPayload,
   TargetBlockPayload,
 } from '../indexer/events';
+
+const DEFAULT_TIMEOUT = 900000;
 
 @Injectable()
 export class HealthService {
@@ -17,6 +20,14 @@ export class HealthService {
   private currentProcessingHeight?: number;
   private currentProcessingTimestamp?: number;
   private blockTime = 6000;
+  private healthTimeout: number;
+
+  constructor(protected nodeConfig: NodeConfig) {
+    this.healthTimeout = Math.max(
+      DEFAULT_TIMEOUT,
+      this.nodeConfig.timeout * 1000,
+    );
+  }
 
   @OnEvent(IndexerEvent.BlockTarget)
   handleTargetBlock(blockPayload: TargetBlockPayload) {
@@ -48,7 +59,7 @@ export class HealthService {
     }
     if (
       !this.currentProcessingTimestamp ||
-      Date.now() - this.currentProcessingTimestamp > this.blockTime * 10
+      Date.now() - this.currentProcessingTimestamp > this.healthTimeout
     ) {
       throw new Error('Indexer is not healthy');
     }